chore: remove mattermost.js and use index.ts as app entry
