+++ index.ts
@@ -1,11 +1,18 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import {Platform} from 'react-native';
+import {DeviceEventEmitter, Platform} from 'react-native';
 import 'react-native-gesture-handler';
+import {ComponentDidAppearEvent, ComponentDidDisappearEvent, Navigation} from 'react-native-navigation';
 
+import {Navigation as NavigationConstants, Screens} from './app/constants';
+import GlobalEventHandler from './app/init/global_event_handler';
+import './app/init/fetch';
+import {initialLaunch} from './app/init/launch';
+import ManagedApp from './app/init/managed_app';
+import {registerScreens} from './app/screens/index';
+import EphemeralStore from './app/store/ephemeral_store';
 import setFontFamily from './app/utils/font_family';
-import './app/mattermost';
 
 declare const global: { HermesInternal: null | {} };
 
@@ -33,4 +40,41 @@ if (Platform.OS === 'android') {
     const ShareExtension = require('share_extension/index.tsx').default;
     const AppRegistry = require('react-native/Libraries/ReactNative/AppRegistry');
     AppRegistry.registerComponent('MattermostShare', () => ShareExtension);
+}
+
+Navigation.events().registerAppLaunchedListener(() => {
+    GlobalEventHandler.init();
+    ManagedApp.init();
+    registerNavigationListeners();
+    registerScreens();
+    initialLaunch();
+});
+
+const registerNavigationListeners = () => {
+    Navigation.events().registerComponentDidAppearListener(componentDidAppearListener);
+    Navigation.events().registerComponentDidDisappearListener(componentDidDisappearListener);
+};
+
+function componentDidAppearListener({componentId}: ComponentDidAppearEvent) {
+    EphemeralStore.addNavigationComponentId(componentId);
+
+    switch (componentId) {
+        case 'MainSidebar':
+            DeviceEventEmitter.emit(NavigationConstants.MAIN_SIDEBAR_DID_OPEN, this.handleSidebarDidOpen);
+            DeviceEventEmitter.emit(NavigationConstants.BLUR_POST_DRAFT);
+            break;
+        case 'SettingsSidebar':
+            DeviceEventEmitter.emit(NavigationConstants.BLUR_POST_DRAFT);
+            break;
+    }
+}
+
+function componentDidDisappearListener({componentId}: ComponentDidDisappearEvent) {
+    if (componentId !== Screens.CHANNEL) {
+        EphemeralStore.removeNavigationComponentId(componentId);
+    }
+
+    if (componentId === 'MainSidebar') {
+        DeviceEventEmitter.emit(NavigationConstants.MAIN_SIDEBAR_DID_CLOSE);
+    }
 }
